My_chatbot/requirements.txt
python-telegram-bot==13.15
gpt4all
